+++ cl_manycore/libraries/bsg_manycore_responder.h
@@ -1,0 +1,4 @@
+#ifndef BSG_MANYCORE_RESPONDER_H
+#define BSG_MANYCORE_RESPONDER_H
+
+#endif
